requirements.txt
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==3.0.0
Flask-Session==0.5.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
protobuf==4.21.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
